+++ terraform/azdoit-test/main.tf
@@ -138,9 +138,9 @@ resource "azurerm_linux_virtual_machine" "test" {
   }
 
   source_image_reference {
-    publisher = "Canonical"
-    offer     = "0001-com-ubuntu-server-jammy"
-    sku       = "22_04-lts-gen2"
+    publisher = "MicrosoftCBLMariner"
+    offer     = "azure-linux-4"
+    sku       = "azure-linux-4-gen2"
     version   = "latest"
   }